Make it possible to assign int, etc. to StrictInt, PositiveInt etc.
Currently, when the `init_typed = True` mypy plugin option is used, I
get many mypy errors such as

    error: Argument "foo" to "Foo" has incompatible type "int"; expected "StrictInt"  [arg-type]

This happens because it is not allow to assign a base class (`int` in
this example) to a subclass (`StrictInt` in this example).

This problem was previously fixed only for `StrictBool`, by making it
the same as `bool` when `TYPE_CHECKING`. Do the same for all the other
relevant types.